OSIS-5846 : Refactor
