+++ ddd/logic/learning_unit/domain/service/create_effective_class.py
@@ -23,14 +23,17 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+from decimal import Decimal
 from typing import List
 
+from base.ddd.utils.business_validator import MultipleBusinessExceptions
 from ddd.logic.learning_unit.builder.effective_class_builder import EffectiveClassBuilder
 from ddd.logic.learning_unit.commands import CreateEffectiveClassCommand
 from ddd.logic.learning_unit.domain.model.effective_class import EffectiveClassIdentity
 from ddd.logic.learning_unit.domain.model.learning_unit import LearningUnit
 from ddd.logic.learning_unit.domain.service.can_access_creation_effective_class import CanAccessCreationEffectiveClass
-from ddd.logic.learning_unit.domain.service.can_create_effective_class import CanCreateEffectiveClass
+from ddd.logic.learning_unit.domain.validator.exceptions import CodeClassAlreadyExistForUeException, \
+    AnnualVolumeInvalidException
 from ddd.logic.learning_unit.repository.i_learning_unit import ILearningUnitRepository
 from osis_common.ddd import interface
 
@@ -49,13 +52,51 @@ class CreateEffectiveClass(interface.DomainService):
             learning_unit=learning_unit,
             learning_unit_repository=learning_unit_repository
         )
-        CanCreateEffectiveClass().check(
-            learning_unit=learning_unit,
-            all_existing_class_identities=all_existing_class_identities,
+
+        cls.can_create_effective_class(
             cmd=cmd,
+            all_existing_class_identities=all_existing_class_identities,
+            learning_unit=learning_unit
         )
+
         effective_class = EffectiveClassBuilder.build_from_command(
             cmd=cmd,
             learning_unit=learning_unit
         )
         return effective_class
+
+    @classmethod
+    def can_create_effective_class(
+            cls,
+            cmd: 'CreateEffectiveClassCommand',
+            all_existing_class_identities: List['EffectiveClassIdentity'],
+            learning_unit: 'LearningUnit'
+    ):
+        exceptions = set()  # type Set[BusinessException]
+
+        if _is_effective_class_volumes_inconsistent_with_learning_unit_volume_annual(learning_unit, cmd):
+            exceptions.add(AnnualVolumeInvalidException(learning_unit))
+
+        if all_existing_class_identities:
+            for class_id in all_existing_class_identities:
+                if class_id.learning_unit_identity == learning_unit.entity_id and class_id.class_code == cmd.class_code:
+                    exceptions.add(CodeClassAlreadyExistForUeException(learning_unit.entity_id, cmd.class_code))
+
+        if exceptions:
+            raise MultipleBusinessExceptions(exceptions=exceptions)
+
+
+def _is_effective_class_volumes_inconsistent_with_learning_unit_volume_annual(
+        learning_unit: 'LearningUnit',
+        cmd: 'CreateEffectiveClassCommand'
+) -> Decimal:
+    practical_part = learning_unit.practical_part
+    lecturing_part = learning_unit.lecturing_part
+
+    if practical_part and not lecturing_part:
+        volume_annual = practical_part.volumes.volume_annual
+    else:
+        volume_annual = lecturing_part.volumes.volume_annual
+
+    sum_q1_q2 = cmd.volume_first_quadrimester + cmd.volume_second_quadrimester
+    return volume_annual <= 0 or sum_q1_q2 != volume_annual
